requirements.txt
requests
urllib3
python_dotenv
selenium==3.141.0